Round EstimatedDiameter and MissDistance to 3 decimal places

+++ NasaApiBackend/Models/Neo/OutputNeoModel.cs
@@ -4,9 +4,9 @@
     {
         public long Id { get; set; }
         public string Name { get; set; }
-        public double EstimatedDiameter { get; set; }
+        public decimal EstimatedDiameter { get; set; }
         public bool IsPotentiallyHazardousAsteroid { get; set; }
         public DateTime CloseApproachDateFull { get; set; }
-        public double MissDistance { get; set; }
+        public decimal MissDistance { get; set; }
     }
 }
